🚚 reorganized SidenavComponent into a single directory
