+++ libs/ui/layout/src/lib/components/index.ts
@@ -1,8 +1,5 @@
 export { FooterContainerComponent, FooterComponent } from './footer';
 export { ToolbarContainerComponent, ToolbarComponent } from './toolbar';
+export { SidenavContainerComponent, SidenavComponent } from './sidenav';
 
 export { ShellComponent } from './shell/shell.component';
-
-export { SidenavContainerComponent } from './sidenav/sidenav-container/sidenav-container.component';
-export { SidenavComponent } from './sidenav/sidenav-container/sidenav/sidenav.component';
-
