+++ libs/ui/layout/src/lib/components/sidenav/component/sidenav-container/sidenav/sidenav.component.ts
@@ -1,0 +1,9 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  selector: 'ia-layout-sidenav',
+  templateUrl: './sidenav.component.html',
+  styleUrls: ['./sidenav.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class SidenavComponent { }
