+++ libs/ui/layout/src/lib/components/sidenav/index.ts
@@ -1,0 +1,8 @@
+export type { } from './model/sidenav.model';
+
+export { } from './constant/sidenav.constants';
+
+export { SidenavContainerComponent } from './component/sidenav-container/sidenav-container.component';
+export { SidenavComponent } from './component/sidenav-container/sidenav/sidenav.component';
+
+export { } from './service/sidenav.service';
