+++ libs/ui/layout/src/lib/components/sidenav/service/sidenav.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { SidenavService } from './sidenav.service';
+
+describe('SidenavService', () => {
+  let service: SidenavService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(SidenavService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
