+++ libs/ui/layout/src/lib/components/sidenav/service/sidenav.service.ts
@@ -1,0 +1,6 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class SidenavService { }
