--- libs/ui/layout/src/lib/components/sidenav/sidenav-container/sidenav/sidenav.component.ts
@@ -1,11 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'ia-layout-sidenav',
-  templateUrl: './sidenav.component.html',
-  styleUrls: ['./sidenav.component.scss']
-})
-export class SidenavComponent implements OnInit {
-  constructor() { }
-  ngOnInit(): void { }
-}
