+++ libs/ui/layout/src/lib/models/index.ts
@@ -12,8 +12,8 @@ export type {
   ToolbarDataModel
 } from '../components/toolbar';
 
-export type { NavigationItemModel } from './navigation/navigation.model';
+export type { } from '../components/sidenav';
 
-export { } from './sidenav/sidenav.model';
+export type { NavigationItemModel } from './navigation/navigation.model';
 
 export type { LayoutDataModel } from './layout/layout.model';
